+++ service/src/test/java/bio/terra/workspace/app/configuration/external/controller/WorkspaceApiControllerTest.java
@@ -1,12 +1,12 @@
 package bio.terra.workspace.app.configuration.external.controller;
 
-import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.createDefaultWorkspace;
 import static bio.terra.workspace.common.utils.MockMvcUtils.CLONE_WORKSPACE_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockMvcUtils.UPDATE_WORKSPACES_V1_PROPERTIES_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockMvcUtils.WORKSPACES_V1_BY_UUID_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockMvcUtils.WORKSPACES_V1_PATH;
 import static bio.terra.workspace.common.utils.MockMvcUtils.addAuth;
 import static bio.terra.workspace.common.utils.MockMvcUtils.addJsonContentType;
+import static bio.terra.workspace.common.utils.MockMvcUtils.createWorkspace;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -146,7 +146,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
 
   @Test
   public void updateWorkspace() throws Exception {
-    ApiCreatedWorkspace workspace = createDefaultWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace workspace = createWorkspace(mockMvc, objectMapper);
     String newDisplayName = "new workspace display name";
     String newUserFacingId = "new-ufid";
     String newDescription = "new description for the workspace";
@@ -213,7 +213,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
 
   @Test
   public void deleteWorkspaceProperties() throws Exception {
-    UUID workspaceId = createDefaultWorkspace(mockMvc, objectMapper).getId();
+    UUID workspaceId = createWorkspace(mockMvc, objectMapper).getId();
     ApiWorkspaceDescription sourceWorkspace = getWorkspaceDescription(workspaceId);
     ArrayList propertyUpdate = new ArrayList<>(Arrays.asList("foo", "foo1"));
 
@@ -236,7 +236,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
 
   @Test
   public void updateWorkspaceProperties() throws Exception {
-    UUID workspaceId = createDefaultWorkspace(mockMvc, objectMapper).getId();
+    UUID workspaceId = createWorkspace(mockMvc, objectMapper).getId();
     ApiWorkspaceDescription sourceWorkspace = getWorkspaceDescription(workspaceId);
     Map<String, String> properties = Map.of("foo", "bar");
 
@@ -259,7 +259,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
 
   @Test
   public void cloneWorkspace() throws Exception {
-    UUID workspaceId = createDefaultWorkspace(mockMvc, objectMapper).getId();
+    UUID workspaceId = createWorkspace(mockMvc, objectMapper).getId();
     ApiWorkspaceDescription sourceWorkspace = getWorkspaceDescription(workspaceId);
 
     String serializedGetResponse =
@@ -344,7 +344,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
   @Test
   public void getWorkspaceIncludesPolicy() throws Exception {
     // No need to actually pass policy inputs because TPS is mocked.
-    ApiCreatedWorkspace workspace = createDefaultWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace workspace = createWorkspace(mockMvc, objectMapper);
 
     ApiTpsPaoGetResult getPolicyResult =
         emptyWorkspacePao()
@@ -362,7 +362,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
   @Test
   public void tpsDisabledGetWorkspaceExcludesPolicy() throws Exception {
     when(mockFeatureConfiguration.isTpsEnabled()).thenReturn(false);
-    ApiCreatedWorkspace workspace = createDefaultWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace workspace = createWorkspace(mockMvc, objectMapper);
 
     ApiWorkspaceDescription gotWorkspace = getWorkspaceDescription(workspace.getId());
     assertNull(gotWorkspace.getPolicies());
@@ -371,8 +371,8 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
   @Test
   public void listWorkspaceIncludesPolicy() throws Exception {
     // No need to actually pass policy inputs because TPS is mocked.
-    ApiCreatedWorkspace workspace = createDefaultWorkspace(mockMvc, objectMapper);
-    ApiCreatedWorkspace noPolicyWorkspace = createDefaultWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace workspace = createWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace noPolicyWorkspace = createWorkspace(mockMvc, objectMapper);
     when(mockSamService.listWorkspaceIdsAndHighestRoles(any(), any()))
         .thenReturn(
             ImmutableMap.of(
@@ -405,7 +405,7 @@ public class WorkspaceApiControllerTest extends BaseUnitTest {
   @Test
   public void tpsDisabledListWorkspaceExcludesPolicy() throws Exception {
     when(mockFeatureConfiguration.isTpsEnabled()).thenReturn(false);
-    ApiCreatedWorkspace workspace = createDefaultWorkspace(mockMvc, objectMapper);
+    ApiCreatedWorkspace workspace = createWorkspace(mockMvc, objectMapper);
     when(mockSamService.listWorkspaceIdsAndHighestRoles(any(), any()))
         .thenReturn(ImmutableMap.of(workspace.getId(), WsmIamRole.OWNER));
